requirements.txt
fastapi
uvicorn
qrcode[pil]
python-multipart
psycopg2-binary